seperate components to different files and add thumbnail instead of video
