+++ package.json
@@ -24,6 +24,7 @@
     "react-native-reanimated": "^3.3.0",
     "react-native-safe-area-context": "^4.5.3",
     "react-native-screens": "^3.20.0",
+    "react-native-thumbnail-video": "^0.1.2",
     "react-native-video": "^5.2.1"
   },
   "devDependencies": {
